Refactored username sanitization upgrade migration to run in job batches of 1000 users to prevent server resource overloading
