+++ upgrades/2022_04_19_164837_sanitize_usernames.php
@@ -1,19 +1,11 @@
 <?php
 
-use ProcessMaker\Models\User;
 use Illuminate\Support\Facades\DB;
-use Illuminate\Support\Facades\Validator;
+use ProcessMaker\Jobs\SanitizeUsernames as SanitizeUsernamesJob;
 use ProcessMaker\Upgrades\UpgradeMigration as Upgrade;
 
 class SanitizeUsernames extends Upgrade
 {
-    /**
-     * The version of ProcessMaker being upgraded *to*
-     *
-     * @var string example: 4.2.28
-     */
-    public $to = '4.2.30-RC1';
-
     /**
      * Run any validations/pre-run checks to ensure the environment, settings,
      * packages installed, etc. are right correct to run this upgrade.
@@ -45,32 +37,11 @@ class SanitizeUsernames extends Upgrade
      */
     public function up()
     {
-        $users = DB::table('users')->select(['id','username'])
-                                   ->orderBy('id')
-                                   ->get();
-
-        foreach ($users as $user) {
-
-            // Store the pre-update username for comparison
-            $pre_update_username = $user->username;
-            $updated_username = static::filterAndValidateUsername($user->username, $user->id);
-
-            // If it's the same, then it was already valid
-            if ($pre_update_username === $updated_username) {
-                continue;
+        DB::table('users')->select(['id','username'])->orderBy('id')->chunk(1000,
+            static function ($users) {
+                dispatch(new SanitizeUsernamesJob($users))->onQueue('high');
             }
-
-            // Set the valid username to the user
-            DB::table('users')->where('id', '=', $user->id)
-                                     ->update(['username' => $updated_username]);
-
-            // Log the result and continue
-            logger()->info("Username Updated From ({$pre_update_username}) to ({$user->username})", [
-                'user_id' => $user->id,
-                'updated_from_username' => $pre_update_username,
-                'updated_to_username' => $user->username
-            ]);
-        }
+        );
     }
 
 
@@ -82,55 +53,6 @@ class SanitizeUsernames extends Upgrade
     public function down()
     {
         //
-    }
-
-    /**
-     * Provide the user id or User model and receive a validated username string
-     *
-     * @param  string  $username
-     * @param  int  $id
-     *
-     * @return string
-     */
-    public static function filterAndValidateUsername(string $username, int $id): string
-    {
-        $generator = static function (int $i = null) use (&$username) {
-            preg_match_all('/[^a-zA-Z\d\s_-]/m', $username, $invalid_chars, PREG_SET_ORDER, 0);
-
-            $invalid_chars = collect($invalid_chars)->flatten()->unique()->values();
-            $username = ! blank($invalid_chars) ? str_replace($invalid_chars->all(), '', $username) : $username;
-
-            if (is_int($i) && $i !== 0) {
-                $username .= $i;
-            }
-        };
-
-        // Will validate/filter the username
-        // and set $valid_username
-        $generator($i = 0);
-
-        // Ensure uniqueness for the username
-        $unique_username_query =  DB::table('users')->where('username', '=', $username)
-                                                    ->where('id', '!=', $id)
-                                                    ->orderBy('id');
-
-        // if it's not unique, append an index number to it
-        while ($unique_username_query->exists()) {
-            $generator(++$i);
-        }
-
-        // Grab the User model rules
-//        $rules = (object) User::rules($username);
-
-        // Make sure the new, filtered username is valid
-        // with the User-model provided rules
-//        Validator::make($rules->username,
-//            ['username' => $username]
-//        )->validate();
-
-        // Once we know it's a unique, valid
-        // username, send it back
-        return $username;
     }
 }
